feat:adds routing in main file
