+++ client/src/Components/Background/Background.css
@@ -1,0 +1,18 @@
+.background{
+    height: 100%;
+    width: 100%;
+    float:left;
+    top:0;
+    left:0;
+    right:0;
+    bottom:0;
+    padding:none;
+    position:fixed;
+    object-fit:cover;
+    z-index:-1;
+}
+
+
+.fade-in{
+    animation: fadeIn 2500ms ease-in-out;
+}
