--- client/src/Components/Hero.css
@@ -1,61 +1,0 @@
-.hero{
-    margin:0px 60px;
-    margin-top: 135px;
-}
-.hero-text{
-    color:#fff;
-    font-size: 95px;
-    font-weight: 500;
-    line-height: 200px;
-}
-.hero-explore{
-    display:flex;
-    align-items:center;
-    gap:25px;
-    height: 50px;
-    width:fit-content;
-    margin: top 70px;
-    padding:10px 8px;
-    padding-left:30px;
-    border-radius:60px;
-    background-color:white;
-    cursor: pointer;
-}
-.arrow-icon{
-    width: 40px;
-    padding-right: 10px;
-    border-radius:35px;
-}
-.hero-dot-play{
-    margin-top:120px;
-    display: flex;
-    justify-content: space-between;
-
-}
-.hero-dots{
-    display:flex;
-    align-items:center;
-    gap:23px;
-    list-style:none;
-}
-.hero-dot{
-    width: 15px;
-    height: 15px;
-    background:#fff ;
-    border-radius: 7.5px;
-    cursor: pointer;
-
-}
-.hero-dot.orange{
-    background: orangered;
-}
-.hero-play{
-    display: flex;
-    align-items: center;
-    gap:30px;
-
-}
-.hero-play{
-    color: #fff;
-    font-size:25px;
-}
