+++ client/src/Components/LoginSignup/LoginForm.jsx
@@ -1,0 +1,44 @@
+import React from 'react';
+import { Formik, Form, Field, ErrorMessage } from 'formik';
+import * as Yup from 'yup';
+import './LoginSignup.css';
+
+const LoginForm = () => {
+  const initialValues = {
+    email: '',
+    password: '',
+  };
+
+  const validationSchema = Yup.object({
+    email: Yup.string().email('Invalid email format').required('Required'),
+    password: Yup.string().min(6, 'Password must be at least 6 characters').required('Required'),
+  });
+
+  const onSubmit = (values, { resetForm }) => {
+    console.log(values);
+    resetForm();
+  };
+
+  return (
+    <div className="container">
+      <h1>Login</h1>
+      <Formik initialValues={initialValues} validationSchema={validationSchema} onSubmit={onSubmit}>
+        <Form>
+          <div className="form-group">
+            <label htmlFor="email">Email</label>
+            <Field type="email" id="email" name="email" />
+            <ErrorMessage name="email" component="div" className="error" />
+          </div>
+          <div className="form-group">
+            <label htmlFor="password">Password</label>
+            <Field type="password" id="password" name="password" />
+            <ErrorMessage name="password" component="div" className="error" />
+          </div>
+          <button type="submit">Login</button>
+        </Form>
+      </Formik>
+    </div>
+  );
+};
+
+export default LoginForm;
