+++ client/src/Components/LoginSignup/SignupForm.jsx
@@ -1,0 +1,51 @@
+import React from 'react';
+import { Formik, Form, Field, ErrorMessage } from 'formik';
+import * as Yup from 'yup';
+import './LoginSignup.css';
+
+const SignupForm = () => {
+  const initialValues = {
+    username: '',
+    email: '',
+    password: '',
+  };
+
+  const validationSchema = Yup.object({
+    username: Yup.string().required('Username Required'),
+    email: Yup.string().email('Invalid email format').required('Fill in your email'),
+    password: Yup.string().min(6, 'Password must be at least 6 characters').required('Fill in your password'),
+  });
+
+  const onSubmit = (values, { resetForm }) => {
+    console.log(values);
+    resetForm();
+  };
+
+  return (
+    <div className="container">
+      <h1>Signup</h1>
+      <Formik initialValues={initialValues} validationSchema={validationSchema} onSubmit={onSubmit}>
+        <Form>
+          <div className="form-group">
+            <label htmlFor="username">UserName</label>
+            <Field type="text" id="username" name="username" />
+            <ErrorMessage name="username" component="div" className="error" />
+          </div>
+          <div className="form-group">
+            <label htmlFor="email">Email</label>
+            <Field type="email" id="email" name="email" />
+            <ErrorMessage name="email" component="div" className="error" />
+          </div>
+          <div className="form-group">
+            <label htmlFor="password">Password</label>
+            <Field type="password" id="password" name="password" />
+            <ErrorMessage name="password" component="div" className="error" />
+          </div>
+          <button type="submit">Signup</button>
+        </Form>
+      </Formik>
+    </div>
+  );
+};
+
+export default SignupForm;
