--- client/src/Components/Navbar.css
@@ -1,42 +1,0 @@
-.nav {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 10px 20px;
-    background-color: #e2cbcb;
-    color: white;
-  }
-  
-  .nav-logo {
-    font-size: 24px;
-    font-weight: bold;
-    font-family: 'Arial', sans-serif;
-    color: #fff;
-  }
-  
-  .nav-menu {
-    list-style: none;
-    display: flex;
-    margin: 0;
-    padding: 0;
-  }
-  
-  .nav-menu li {
-    margin: 0 15px;
-  }
-  
-  .nav-menu li a {
-    text-decoration: none;
-    color: white;
-    font-size: 16px;
-    transition: color 0.3s;
-  }
-  
-  .nav-menu li a:hover {
-    color: #ff6347; 
-  }
-  
-  .nav-menu li a.active {
-    border-bottom: 2px solid #ff6347;
-  }
-
